Apply rule "Math operands should be cast before assignment"

+++ src/main/java/org/springframework/context/support/JapanDictGui.java
@@ -2608,7 +2608,7 @@ public class JapanDictGui extends JPanel implements ActionListener, Initializing
 		//
 		while (System.currentTimeMillis() - currentTimeMillis < NumberUtils.max(
 				toMillis(instance != null ? instance.storkeImageDuration : null, 20000),
-				StringUtils.length(JapanDictEntry.getText(japanDictEntry)) * 4100, 0)) {
+				((long) StringUtils.length(JapanDictEntry.getText(japanDictEntry))) * 4100, 0)) {
 			//
 			if (before == null) {
 				//
